refactor(sequencer): fix prepare proposal metrics (#1211)
## Summary
The sequencer's `prepare_proposal()` metrics got shifted to
`process_proposal()` across some refactors.

## Changes
Put metrics back in proper places and removed unneeded metric.

## Testing
Ran locally 

## Metrics
Removed the `PREPARE_PROPOSAL_EXCLUDED_TRANSACTIONS_DECODE_FAILURE`
metric as the mempool should be filtering these transactions out.

+++ crates/astria-sequencer/src/metrics.rs
@@ -167,8 +167,10 @@ impl Metrics {
             .increment(1);
     }
 
-    pub(crate) fn set_prepare_proposal_excluded_transactions(&self, count: f64) {
-        self.prepare_proposal_excluded_transactions.set(count);
+    pub(crate) fn set_prepare_proposal_excluded_transactions(&self, count: usize) {
+        #[allow(clippy::cast_precision_loss)]
+        self.prepare_proposal_excluded_transactions
+            .set(count as f64);
     }
 
     pub(crate) fn record_proposal_deposits(&self, count: usize) {
